+++ travel-budget-app/src/ui/ui.py
@@ -1,34 +1,24 @@
-from tkinter import ttk
+from ui.login import Login
+from ui.create_user import CreateUser
 
 class UI:
     def __init__(self, root):
         self._root = root
-        self.login_start()
-        #self.create_user_start()
+        self.window = None
+        self.login_view()
 
-    def login_start(self):
-        heading_label = ttk.Label(master=self._root, text="Login")
+    def hide_current_window(self):
+        if self.window:
+            self.window.destroy()
+            #?
+            self.window = None
 
-        username_label = ttk.Label(master=self._root, text="Username")
-        username_entry = ttk.Entry(master=self._root)
+    def login_view(self):
+        self.hide_current_window()
+        self.window = Login(self._root, self.create_user_view)
+        self.window.pack()
 
-        password_label = ttk.Label(master=self._root, text="Password")
-        password_entry = ttk.Entry(master=self._root)
-
-        button_login = ttk.Button(master=self._root, text="Login")
-        button_create_user = ttk.Button(master=self._root, text="Create a new user")
-
-        heading_label.grid(row=0, column=0, columnspan=2)
-
-        username_label.grid(row=1, column=0)
-        username_entry.grid(row=1, column=1)
-
-        password_label.grid(row=2, column=0)
-        password_entry.grid(row=2, column=1)
-
-        button_login.grid(row=3, column=0, columnspan=2)
-        button_create_user.grid(row=4, column=0, columnspan=2)
-
-    def create_user_start(self):
-        pass
-        
+    def create_user_view(self):
+        self.hide_current_window()
+        self.window = CreateUser(self._root, self.login_view)
+        self.window.pack()