+++ travel-budget-app/documentation/changelog.md
@@ -1,6 +1,6 @@
 # Changes
 ## Week 3
 - project initialized
-- UI: login view
-- sqlite database for username and password created, database connection done
+- UI: login and create user view
+- sqlite database for username and password created
 - adding and finding user (username and password) in database tested